+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Set the page title in index.html and upon route changes
 - Refactored the asset name rendering to distinguish page title and asset name decorators
 - Set the route title for the asset and manage plugins pages
+- Change the archived asset name icon to use a unicode icon that can also be rendered as the `document.title`
 
 ## [0.6.2] - 2022-09-03